fix(app): extract pure prompt-equality helpers so prompt-input tests don't load @solidjs/router
prompt.tsx imports useParams from @solidjs/router at module top. That router
module calls a client-only solid-js API at eval time, which throws in bun's
server-side test env (happydom.ts only adds DOM globals). So any test that
VALUE-imports prompt.tsx without mocking @solidjs/router fails with the
misleading "Export named 'isPromptEqual' not found in module prompt.tsx".
command-prepend / command-backspace / command-paste / portable-draft /
path-b-integration all hit this; they only passed in CI because an earlier
test leaked a non-restoring mock.module("@solidjs/router") globally (the
#1084 pattern), so isolated/local runs failed deterministically.

Fix: extract the pure equality/default-prompt helpers (DEFAULT_PROMPT,
isSelectionEqual, isCommandMetaEqual, isPartEqual, isPromptEqual,
isStructurallyEmpty) from prompt.tsx into a new pure module
context/prompt-equality.ts whose imports are all type-only (zero runtime
deps, no @solidjs/router). prompt.tsx imports them back and re-exports them,
so the @/context/prompt public surface is unchanged. All pure-symbol VALUE
consumers (command-prepend.ts, command-backspace.ts, command-paste.ts,
editor-input.ts, editor-serialize.ts, popover-controllers.ts,
homepage-migration.ts + the 6 test files) are redirected to import the values
from @/context/prompt-equality, keeping their type-only imports from
@/context/prompt. prompt-input.tsx is intentionally left as-is (it imports
usePrompt, so it loads the context anyway).

Verification: typecheck 8/8, eslint clean (production files), full app unit
suite 1715 pass / 0 fail (previously 2 deterministic failures), all 5
formerly-broken tests now pass in isolation, codex review GATE PASS.

prompt.tsx 373 -> 319; prompt-equality.ts +58.
