+++ packages/app/src/components/prompt-input/command-backspace.test.ts
@@ -6,7 +6,7 @@
 
 import { describe, expect, test } from "bun:test"
 import type { Prompt, TextPart, FileAttachmentPart, AgentPart, ImageAttachmentPart } from "@/context/prompt"
-import { DEFAULT_PROMPT } from "@/context/prompt"
+import { DEFAULT_PROMPT } from "@/context/prompt-equality"
 import { computeCommandBackspaceResult } from "./command-backspace"
 
 // Helpers -----------------------------------------------------------------
